+++ packages/harpguru-core/src/components/menu-of-scales/hooks/use-scales-itemsz/index.ts
@@ -1,1 +1,0 @@
-export { useScalesItemsz } from './use-scales-itemsz'