+++ packages/harpguru-core/src/components/menu-of-scales/hooks/use-scales-titles/index.ts
@@ -1,1 +1,0 @@
-export { useScalesTitles } from './use-scales-titles'