+++ packages/harpguru-core/src/components/menu-of-scales/menu-of-scales.tsx
@@ -9,35 +9,42 @@ import { Menu } from '../menu'
 import { MenuProps } from '../../types'
 import { colors, getSizes } from '../../styles'
 
-import { getTogglesForDispatcher } from './utils'
-import { useScalesTitles, useScalesItemsz } from './hooks'
+import {
+  getNewBufferedTogglesByTargetScaleForDispatcher,
+  getNewActiveQuizDegreesByToggleForDispatcher,
+} from './utils'
+import {
+  useScaleTitle,
+  useScaleItems,
+  useQuizQuestionTitle,
+  useQuizQuestionItems,
+} from './hooks'
 import { useImmediatelyFlushToggles } from './hooks'
 
 export const MenuOfScales = (menuProps: MenuProps): React.ReactElement => {
-  const { useScalesTitle, useChordsTitle } = useScalesTitles()
-  const { useScalesItems, useChordsItems } = useScalesItemsz()
   useImmediatelyFlushToggles({
     isMenuStashed: menuProps.isMenuStashed,
   })
-  const itemTapHandler = useCallback(useDispatch(getTogglesForDispatcher), [
-    useDispatch,
-    getTogglesForDispatcher,
-  ])
-
-  const useScalesTitleMemo = useCallback(() => useScalesTitle(useGlobal), [
-    useGlobal,
-  ])
-  const useChordsTitleMemo = useCallback(() => useChordsTitle(useGlobal), [
-    useGlobal,
-  ])
-
-  const useScalesItemsMemo = useCallback(
-    () => useScalesItems(useGlobal, itemTapHandler),
-    [useGlobal, itemTapHandler]
+  const scaleItemTapHandler = useCallback(
+    useDispatch(getNewBufferedTogglesByTargetScaleForDispatcher),
+    [useDispatch, getNewBufferedTogglesByTargetScaleForDispatcher]
   )
-  const useChordsItemsMemo = useCallback(
-    () => useChordsItems(useGlobal, itemTapHandler),
-    [useGlobal, itemTapHandler]
+  const useScalesTitleMemo = useCallback(() => useScaleTitle(useGlobal), [
+    useGlobal,
+  ])
+  const useScalesItemsMemo = useCallback(
+    () => useScaleItems(useGlobal, scaleItemTapHandler),
+    [useGlobal, scaleItemTapHandler]
+  )
+
+  const useQuizQuestionTitleMemo = useCallback(() => useQuizQuestionTitle(), [])
+  const quizQuestionTapHandler = useCallback(
+    useDispatch(getNewActiveQuizDegreesByToggleForDispatcher),
+    [useDispatch, getNewActiveQuizDegreesByToggleForDispatcher]
+  )
+  const useQuizQuestionItemsMemo = useCallback(
+    () => useQuizQuestionItems(useGlobal, quizQuestionTapHandler),
+    [useGlobal, quizQuestionTapHandler]
   )
 
   const optionStackPropsz = [
@@ -47,8 +54,8 @@ export const MenuOfScales = (menuProps: MenuProps): React.ReactElement => {
       twoColumns: false,
     },
     {
-      useTitle: useChordsTitleMemo,
-      useItems: useChordsItemsMemo,
+      useTitle: useQuizQuestionTitleMemo,
+      useItems: useQuizQuestionItemsMemo,
       twoColumns: false,
     },
   ]