+++ packages/harpguru-core/src/components/menu-of-scales/utils/get-new-active-quiz-degrees-by-toggle-for-dispatcher/get-new-active-quiz-degrees-by-toggle-for-dispatcher.test.ts
@@ -2,7 +2,7 @@ import { DegreeIds } from 'harpparts'
 
 import type { GlobalState } from '../../../../types'
 
-import { getToggledActiveQuizDegrees } from './get-toggled-active-quiz-degrees'
+import { getNewActiveQuizDegreesByToggleForDispatcher } from './get-new-active-quiz-degrees-by-toggle-for-dispatcher'
 
 test('toggling a degree in a list containing only that degree returns an empty list', () => {
   const inputGlobal = {
@@ -12,7 +12,11 @@ test('toggling a degree in a list containing only that degree returns an empty l
 
   const {
     activeQuizDegrees: newActiveQuizDegrees,
-  } = getToggledActiveQuizDegrees(inputGlobal, unusedDispatcher, DegreeIds.Root)
+  } = getNewActiveQuizDegreesByToggleForDispatcher(
+    inputGlobal,
+    unusedDispatcher,
+    DegreeIds.Root
+  )
 
   expect(newActiveQuizDegrees).toStrictEqual([])
 })
@@ -25,7 +29,11 @@ test('toggling a degree in an empty list returns a list with just that degree in
 
   const {
     activeQuizDegrees: newActiveQuizDegrees,
-  } = getToggledActiveQuizDegrees(inputGlobal, unusedDispatcher, DegreeIds.Root)
+  } = getNewActiveQuizDegreesByToggleForDispatcher(
+    inputGlobal,
+    unusedDispatcher,
+    DegreeIds.Root
+  )
 
   expect(newActiveQuizDegrees).toStrictEqual([DegreeIds.Root])
 })