+++ packages/harpguru-core/src/components/menu-of-scales/utils/get-new-active-quiz-degrees-by-toggle-for-dispatcher/index.ts
@@ -1,0 +1,1 @@
+export { getNewActiveQuizDegreesByToggleForDispatcher } from './get-new-active-quiz-degrees-by-toggle-for-dispatcher'