+++ packages/harpguru-core/src/components/menu-of-scales/utils/get-toggles-for-dispatcher/index.ts
@@ -1,1 +1,0 @@
-export { getTogglesForDispatcher } from './get-toggles-for-dispatcher'