+++ packages/harpguru-core/src/components/menu-of-tunings/menu-of-tunings.tsx
@@ -12,18 +12,15 @@ import { colors, getSizes } from '../../styles'
 import {
   getNewHarpStrataByTuningForDispatcher,
   getNewHarpStrataByValvingForDispatcher,
-  getToggledActiveQuizDegrees,
 } from './utils'
 import {
-  useQuizQuestionTitle,
-  useQuizQuestionItems,
   useTuningTitle,
   useTuningItems,
   useValvingTitle,
   useValvingItems,
 } from './hooks'
 
-export const MenuOfSettings = (menuProps: MenuProps): React.ReactElement => {
+export const MenuOfTunings = (menuProps: MenuProps): React.ReactElement => {
   const useTuningTitleMemo = useCallback(() => useTuningTitle(useGlobal), [
     useGlobal,
   ])
@@ -48,16 +45,6 @@ export const MenuOfSettings = (menuProps: MenuProps): React.ReactElement => {
     [useGlobal, valvingItemTapHandler]
   )
 
-  const useQuizQuestionTitleMemo = useCallback(() => useQuizQuestionTitle(), [])
-  const quizQuestionTapHandler = useCallback(
-    useDispatch(getToggledActiveQuizDegrees),
-    [useDispatch, getToggledActiveQuizDegrees]
-  )
-  const useQuizQuestionItemsMemo = useCallback(
-    () => useQuizQuestionItems(useGlobal, quizQuestionTapHandler),
-    [useGlobal, quizQuestionTapHandler]
-  )
-
   const optionStackPropsz = [
     {
       useTitle: useTuningTitleMemo,
@@ -67,11 +54,6 @@ export const MenuOfSettings = (menuProps: MenuProps): React.ReactElement => {
     {
       useTitle: useValvingTitleMemo,
       useItems: useValvingItemsMemo,
-      twoColumns: false,
-    },
-    {
-      useTitle: useQuizQuestionTitleMemo,
-      useItems: useQuizQuestionItemsMemo,
       twoColumns: false,
     },
   ]