+++ packages/harpguru-core/src/components/menu-of-tunings/utils/index.ts
@@ -1,3 +1,2 @@
 export { getNewHarpStrataByTuningForDispatcher } from './get-new-harpstrata-by-tuning-for-dispatcher'
 export { getNewHarpStrataByValvingForDispatcher } from './get-new-harpstrata-by-valving-for-dispatcher'
-export { getToggledActiveQuizDegrees } from './get-toggled-active-quiz-degrees'